[Gibson] Rename feature_generation.py to helpers.py
